+++ packages/react-workspace/nimble-react/src/number-field/index.ts
@@ -1,6 +1,7 @@
-import { NumberField } from '@ni/nimble-components/dist/esm/number-field';
+import { NumberField, numberFieldTag } from '@ni/nimble-components/dist/esm/number-field';
 import { wrap, type EventName } from '../utilities/react-wrapper';
 
+export { numberFieldTag };
 export { type NumberField };
 export const NimbleNumberField = wrap(NumberField, {
     events: {