+++ packages/react-workspace/nimble-react/src/radio/index.ts
@@ -1,5 +1,6 @@
-import { Radio } from '@ni/nimble-components/dist/esm/radio';
+import { Radio, radioTag } from '@ni/nimble-components/dist/esm/radio';
 import { wrap } from '../utilities/react-wrapper';
 
+export { radioTag };
 export { type Radio };
 export const NimbleRadio = wrap(Radio);