+++ packages/react-workspace/nimble-react/src/table-column/number-text/index.ts
@@ -1,5 +1,6 @@
-import { TableColumnNumberText } from '@ni/nimble-components/dist/esm/table-column/number-text';
+import { TableColumnNumberText, tableColumnNumberTextTag } from '@ni/nimble-components/dist/esm/table-column/number-text';
 import { wrap } from '../../utilities/react-wrapper';
 
+export { tableColumnNumberTextTag };
 export { type TableColumnNumberText };
 export const NimbleTableColumnNumberText = wrap(TableColumnNumberText);