+++ packages/react-workspace/nimble-react/src/text-area/index.ts
@@ -1,6 +1,7 @@
-import { TextArea } from '@ni/nimble-components/dist/esm/text-area';
+import { TextArea, textAreaTag } from '@ni/nimble-components/dist/esm/text-area';
 import { wrap, type EventName } from '../utilities/react-wrapper';
 
+export { textAreaTag };
 export { type TextArea };
 export const NimbleTextArea = wrap(TextArea, {
     events: {