+++ packages/react-workspace/react-client-app/src/components/App.tsx
@@ -16,18 +16,18 @@ import { NimbleCheckbox } from '@ni/nimble-react/checkbox';
 import { NimbleRadioGroup } from '@ni/nimble-react/radio-group';
 import { NimbleRadio } from '@ni/nimble-react/radio';
 import { NimbleTextField } from '@ni/nimble-react/text-field';
-import { NimbleDialog, type Dialog, type DialogRef, DialogUserDismissed } from '@ni/nimble-react/dialog';
-import { NimbleDrawer, type Drawer, type DrawerRef, DrawerUserDismissed, DrawerLocation } from '@ni/nimble-react/drawer';
+import { NimbleDialog, type Dialog, DialogUserDismissed, fromDialogRef } from '@ni/nimble-react/dialog';
+import { NimbleDrawer, type Drawer, DrawerUserDismissed, DrawerLocation, fromDrawerRef } from '@ni/nimble-react/drawer';
 import { NimbleMenu } from '@ni/nimble-react/menu';
 import { NimbleMenuItem, type MenuItemChangeEvent } from '@ni/nimble-react/menu-item';
 import { NimbleAnchorMenuItem } from '@ni/nimble-react/anchor-menu-item';
 import { NimbleMenuButton } from '@ni/nimble-react/menu-button';
 import { NimbleIconAdd } from '@ni/nimble-react/icons/add';
-import { NimbleIconCheck } from '@ni/nimble-react/icons/check';
+import { iconCheckTag, NimbleIconCheck } from '@ni/nimble-react/icons/check';
 import { NimbleIconXmarkCheck } from '@ni/nimble-react/icons/xmark-check';
 import { NimbleSpinner } from '@ni/nimble-react/spinner';
 import { NimbleSwitch } from '@ni/nimble-react/switch';
-import { NimbleTable, type Table, type TableRef, type TableRowExpandToggleEvent, type TableRecord, type TableSetRecordHierarchyOptions } from '@ni/nimble-react/table';
+import { NimbleTable, type Table, type TableRowExpandToggleEvent, type TableRecord, type TableSetRecordHierarchyOptions, fromTableRef } from '@ni/nimble-react/table';
 import { NimbleTableColumnText } from '@ni/nimble-react/table-column/text';
 import { NimbleTableColumnAnchor } from '@ni/nimble-react/table-column/anchor';
 import { NimbleTableColumnDateText } from '@ni/nimble-react/table-column/date-text';
@@ -668,9 +668,7 @@ export function App(): React.JSX.Element {
                                 <div className="sub-container">
                                     <div className="container-label">Dialog</div>
                                     <NimbleDialog
-                                        // Note: Generic types such as Dialog require using the workaround ref type
-                                        // See: https://github.com/ni/nimble/issues/2784
-                                        ref={dialogRef as unknown as DialogRef}
+                                        ref={fromDialogRef(dialogRef)}
                                     >
                                         <span slot="title">This is a dialog</span>
                                         <div>It opened when you pushed the button</div>
@@ -691,9 +689,7 @@ export function App(): React.JSX.Element {
                                 <div className="sub-container">
                                     <div className="container-label">Drawer</div>
                                     <NimbleDrawer
-                                        // Note: Generic types such as Drawer require using the workaround ref type
-                                        // See: https://github.com/ni/nimble/issues/2784
-                                        ref={drawerRef as unknown as DrawerRef}
+                                        ref={fromDrawerRef(drawerRef)}
                                         location={drawerLocation}
                                     >
                                         <header>This is a drawer</header>
@@ -931,9 +927,7 @@ export function App(): React.JSX.Element {
                                 <div className="sub-container">
                                     <div className="container-label">Table</div>
                                     <NimbleTable
-                                        // Note: Generic types such as Table require using the workaround ref type
-                                        // See: https://github.com/ni/nimble/issues/2784
-                                        ref={tableRef as unknown as TableRef}
+                                        ref={fromTableRef(tableRef)}
                                         idFieldName="id"
                                         parentIdFieldName="parentId"
                                         selectionMode="multiple"
@@ -970,7 +964,7 @@ export function App(): React.JSX.Element {
                                             <NimbleMappingIcon
                                                 keyValue="success"
                                                 text="Success"
-                                                icon="nimble-icon-check"
+                                                icon={iconCheckTag}
                                                 severity="success"
                                                 textHidden>
                                             </NimbleMappingIcon>
@@ -1028,9 +1022,7 @@ export function App(): React.JSX.Element {
                                 <div className="sub-container">
                                     <div className="container-label">Table with delayed hierarchy</div>
                                     <NimbleTable
-                                        // Note: Generic types such as Table require using the workaround ref type
-                                        // See: https://github.com/ni/nimble/issues/2784
-                                        ref={delayedHierarchyTableRef as unknown as TableRef}
+                                        ref={fromTableRef(delayedHierarchyTableRef)}
                                         idFieldName="id" parentIdFieldName="parentId" selectionMode="multiple"
                                         onRowExpandToggle={onRowExpandToggle}
                                     >